Add files via upload
I'm thrilled to announce that I've successfully completed my first project as a Python Developer Intern at Verve Bridge, working remotely! This milestone marks a significant step in my journey as a developer, and I'm proud of the hard work and dedication that went into achieving it. I'm grateful for the opportunity to work with an amazing team and look forward to continuing to grow and learn.